+++ django_graph_api/tests/test_basic.py
@@ -74,42 +74,16 @@ def test_hero_name_and_episodes(starwars_data):
         'data': {
             'hero': {
                 'name': 'R2-D2',
-                'appears_in': [{
-                    'name': 'A New Hope',
-                    'number': 4
-                }]
+                'appears_in': [
+                    {
+                        'name': 'A New Hope',
+                        'number': 4
+                    },
+                    {
+                        'name': 'The Empire Strikes Back',
+                        'number': 5
+                    },
+                ]
             },
-        }
-    }
-
-
-def test_episodes_and_characters(starwars_data):
-    document = '''
-        {
-            episodes {
-                name
-                number
-                characters {
-                    name
-                }
-            }
-        }
-        '''
-    assert schema.execute(document) == {
-        'data': {
-            'episodes': [
-                {
-                    'name': 'A New Hope',
-                    'number': 4,
-                    'characters': [
-                        {'name': 'Luke Skywalker'},
-                        {'name': 'Darth Vader'},
-                        {'name': 'Han Solo'},
-                        {'name': 'Leia Organa'},
-                        {'name': 'C-3PO'},
-                        {'name': 'R2-D2'},
-                    ]
-                },
-            ]
         }
     }
